doxygen: fix 'dist' when the html docs have not yet been built
The reason why this did not work is that we do not have a fixed list
of html doc files and it would be  pain to maintain.

The 'find' hack gets a list of html files but only when the package
has already been built. Let's complement this hack with the
html/index.html file which act as a build trigger as it is the target name.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>