Drain all events before synchronizing after SYN_DROPPED
The kernel ring buffer drops all events on SYN_DROPPED, but then continues to
fill up again. So by the time we read the events, the kernel's client buffer is
essentially like this:
  SYN_DROPPED, ev1, ev2, ev3, ...., evN

The kernel's device state represents the device after evN, and that is what
the ioctls return. So we can't actually sync while there are events on the
wire because the events represent an earlier state. So simply discard all
events in the kernel buffer, synchronize, and then start processing again. We
lose some granularity but at least the events are correct.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>