tools: add a touchscreen emulator
Very simple emulator of a touchscreen. Requires an existing device created by
e.g. evemu-device and forwards the events from a touchpad device into that
created device, filtering and scaling as needed.

Touchpads don't always have edges that are actually reachable. A --scale
option is available to use only the center 90% of the touchpad or similar.

Note that the aspect ratio of the touchscreen is not preserved, the axis
ranges simply map into the touchscreen's axis ranges.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>